requirements.txt
openpyxl>=3.0.0
dj-database-url== 2.1.0
APScheduler==3.10.4  
Django>=3.2
django-grappelli==4.0.1
xhtml2pdf>=0.2.5
reportlab>=3.5.68
pandas>=1.3.3
gunicorn>=20.0.4 
django-htmx==1.5.0
# WSGI server for running Django in production
psycopg2>=2.9.1          # PostgreSQL database adapter for Django
django-jazzmin
# Optional: Additional security packages
django-environ>=0.8.1    
# For environment variable management
django-extensions>=3.1.3 
# Additional management commands for Django
django-admin-interface==0.9.0


# Optional: Other utilities
whitenoise>=5.3.0       
 # For serving static files in production
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
djangorestframework-simplejwt
djangorestframework
# Below are other potential dependencies captured via pip freeze
 
